Generic Repository - Add method for handle include collections of properties.
Purchase - Add PO, Add Products, Missing Confirm PO & JOIN Product of one elemnt on list.

+++ Source/POS/App.Core/Entities/Product.cs
@@ -11,20 +11,12 @@ namespace App.Core.Entities
         }
 
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string ImagePath { get; set; }
         public string PartNumber { get; set; }
         public int CategoryId { get; set; }
         public int ClasificationId { get; set; }
         public int InventoryId { get; set; }
-
-        //public string Description { get; set; }
-        //public double Price { get; set; }
-        //public string ImagePath { get; set; }
-        //public string PartNumber { get; set; }
-        //public int CategoryId { get; set; }
-        //public int ClasificationId { get; set; }
-
         public virtual Category Category { get; set; }
         public virtual Clasification Clasification { get; set; }
         public virtual Inventory Inventory { get; set; }
